+++ R/fit_growth_gompertz.R
@@ -13,8 +13,8 @@
 #' # Fit the data given in columns Time and OD600
 #' fit_growth_gompertz(mydata, Time, OD600)}
 fit_growth_gompertz <- function(df, time, data, ...) {
-    fit_growth_parametric(df, time = time, data = data,
-                          control = grofit.control(model.type="gompertz"), ...)
+    ctl <- grofit.control(model.type = "gompertz", suppress.messages = TRUE)
+    fit_growth_parametric(df, time = time, data = data, control = ctl, ...)
 }
 
 #' @export
@@ -27,6 +27,7 @@ fit_growth_gompertz <- function(df, time, data, ...) {
 #' fit_growth_gompertz_(df=mydata, time_col='Time', data_col='OD600')}
 #'
 fit_growth_gompertz_ <- function(df, time_col, data_col, ...) {
+    ctl <- grofit.control(model.type = "gompertz", suppress.messages = TRUE)
     fit_growth_parametric_(df = df, time_col = time_col, data_col = data_col,
-                           control = grofit.control(model.type="gompertz"), ...)
+                           control = ctl, ...)
 }